Cleaned up the component structure
